requirements.txt

# Building
wheel
setuptools

# Coverage
coverage

# Sphinx
Sphinx
sphinx-rtd-theme

# Project
imutils
PyQt5
opencv-python-headless
numpy
pyqtgraph
scipy
pySerial